operator: derive the bootstrap env from values via a shared chart helper
Review follow-up: instead of rendering the StatefulSet and fishing the
bootstrap-yaml-envsubst container out of it by name, extract the env
construction into an exported chart helper, BootstrapTemplateEnvVars, and
have both the container and clusterConfigFor use it. The rendered-resource
lookup (and findContainer) goes away entirely; env injected through
PodTemplate overlays is deliberately not mirrored — those target the STS or
job pods, and cluster-config values have these dedicated hooks.

TestPostInstallUpgradeJobIndex's successor becomes a drift guard: the
templater container on the StatefulSet (and the job, when enabled) must
carry exactly the helper's env and no EnvFrom, so an env source added to
the container but not the helper cannot silently fail to resolve in the
operator's in-process reify.

Fold cluster_config_bootstrap_env_test.go into redpanda_controller_test.go.
That file is package redpanda_test while the tests exercise the unexported
clusterConfigFor, so a small ClusterConfigForTesting shim in the existing
common_test.go (package redpanda, same test binary) bridges the gap.
